feat: move /agent to /reverse-recruiter with a dynamic sitemap
/agent now 301s to /reverse-recruiter, which carries the SEO page.
sitemap.xml is generated from code (static pages, blog posts, pSEO
job pages) instead of the stale 5-URL static file, and robots.txt
points at it.
