--- public/sitemap.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset
-      xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
-            http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
-
-<url>
-  <loc>https://resumetailor.ai/</loc>
-  <lastmod>2024-03-31T02:09:28+00:00</lastmod>
-  <priority>1.00</priority>
-</url>
-<url>
-  <loc>https://resumetailor.ai/login</loc>
-  <lastmod>2024-03-31T02:09:28+00:00</lastmod>
-  <priority>0.80</priority>
-</url>
-<url>
-  <loc>https://resumetailor.ai/signup</loc>
-  <lastmod>2024-03-31T02:09:28+00:00</lastmod>
-  <priority>0.80</priority>
-</url>
-<url>
-  <loc>https://resumetailor.ai/forgot-password</loc>
-  <lastmod>2024-03-31T02:09:28+00:00</lastmod>
-  <priority>0.64</priority>
-</url>
-
-<url>
-  <loc>https://resumetailor.ai/ai-resume-builder</loc>
-  <lastmod>2024-03-31T02:09:28+00:00</lastmod>
-  <priority>0.64</priority>
-</url>
-
-
-</urlset>
